Fix the command queue issue mainly on Windows 10.
